requirements.txt
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
streamlit
torch